+ fix warning in HMM_bestpath + perceptron + new classifier hierarchy: classifier -> (LinearClassifier/KernelMachine), LinearClassifier -> Perceptron, KernelMachine ->(SVM/KernelPerceptron/...) + add type LLONG/ULLONG (16 byte integer)

+++ src/lib/Mathmatics.h
@@ -79,7 +79,7 @@ public:
 
 	/// signum of type T variable a 
 	template <class T>
-	static inline REAL sign(REAL a)
+	static inline INT sign(REAL a)
 	{
 		if (a==0)
 			return 0;
